requirements.txt
attrs==23.1.0; python_version >= "3.7" and python_version < "4.0"
aws-cdk-lib==2.76.0; python_version >= "3.7" and python_version < "4.0"
aws-cdk.asset-awscli-v1==2.2.175; python_version >= "3.7" and python_version < "4.0"
aws-cdk.asset-kubectl-v20==2.1.1; python_version >= "3.7" and python_version < "4.0"
aws-cdk.asset-node-proxy-agent-v5==2.0.147; python_version >= "3.7" and python_version < "4.0"
cattrs==22.2.0; python_version >= "3.7" and python_version < "4.0"
constructs==10.2.25; python_version >= "3.7" and python_version < "4.0"
exceptiongroup==1.1.1; python_version >= "3.7" and python_version < "3.11"
importlib-resources==5.12.0; python_version >= "3.7" and python_version < "4.0"
jsii==1.81.0; python_version >= "3.7" and python_version < "4.0"
publication==0.0.3; python_version >= "3.7" and python_version < "4.0"
python-dateutil==2.8.2; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.3.0"
six==1.16.0; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.3.0"
typeguard==2.13.3; python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.5.3"
typing-extensions==4.5.0; python_version >= "3.7" and python_version < "4.0"
